split build.py to create a file containing all the items for the new function`
